requirements.txt
deepl==1.19.1
pyass==0.1.4
srt==3.5.3
python-dotenv==1.0.0
